backend/requirements.txt
fastapi
uvicorn
google-genai
python-dotenv
azure-cognitiveservices-speech
letta-client
openai